requirements.txt
fastapi==0.75.2
marketplace-standard-app-api~=0.4
DLite-Python == 0.3.9
uvicorn<1.0.0
